Extract sqs event validator to separate file
